parte de clientes funcional, muestra mensaje con versión de bootstrap incorrecta

+++ tema-07/proyectos/07-7.3perfiles-gesbank-david/mvc-proyect/views/cuentas/main/index.php
@@ -3,7 +3,7 @@
 
 <head>
     <!-- head -->
-    <?php require_once("template/partials/head.php"); ?>
+    <?php require_once("template/partials/head.php");  ?>
     <title>Cuentas - GESBANK</title>
 </head>
 
@@ -14,10 +14,10 @@
         <?php require_once "template/partials/menuAut.php"; ?>
         <!-- cabecera o titulo -->
         <?php include "views/cuentas/partials/header.php" ?>
-        <!-- Mensajes -->
-        <?php include "template/partials/notify.php" ?>
         <!-- Menu principal -->
         <?php require_once "views/cuentas/partials/menu.php" ?>
+        <!-- Mensaje -->
+        <?php require_once "template/partials/mensaje.php" ?>
         <table class="table">
             <thead>
                 <tr>
@@ -34,46 +34,29 @@
                 </tr>
             </thead>
             <tbody>
-                <?php foreach ($this->cuentas as $cuenta): ?>
+                <?php foreach ($this->cuentas as $cuenta) : ?>
                     <tr>
-                        <td>
-                            <?= $cuenta->id ?>
-                        </td>
-                        <td>
-                            <?= $cuenta->num_cuenta ?>
-                        </td>
-                        <td>
-                            <?= $cuenta->cliente ?>
-                        </td>
-                        <td>
-                            <?= $cuenta->fecha_alta ?>
-                        </td>
-                        <td>
-                            <?= $cuenta->fecha_ul_mov ?>
-                        </td>
-                        <td class="text-end">
-                            <?= number_format($cuenta->num_movtos, 0, ',', '.') ?>
-                        </td>
-                        <td class="text-end">
-                            <?= number_format($cuenta->saldo, 2, ',', '.') ?> €
-                        </td>
+                        <td><?= $cuenta->id ?></td>
+                        <td><?= $cuenta->num_cuenta ?></td>
+                        <td><?= $cuenta->cliente ?></td>
+                        <td><?= $cuenta->fecha_alta ?></td>
+                        <td><?= $cuenta->fecha_ul_mov ?></td>
+                        <td class="text-end"><?= number_format($cuenta->num_movtos, 0, ',', '.')?></td>
+                        <td class="text-end"><?= number_format($cuenta->saldo, 2, ',', '.')?> €</td>
                         <td style="display:flex; justify-content:space-between;">
-                            <a href="<?= URL ?>cuentas/delete/<?= $cuenta->id ?>" title="Eliminar"
-                                onclick="return confirm('Confirmar eliminación Cuenta')"> <i class="bi bi-trash"></i> </a>
-                            <a href="<?= URL ?>cuentas/editar/<?= $cuenta->id ?>" title="Editar"> <i
-                                    class="bi bi-pencil"></i> </a>
-                            <a href="<?= URL ?>cuentas/mostrar/<?= $cuenta->id ?>" title="Mostrar"> <i
-                                    class="bi bi-eye"></i> </a>
-                            <a href="#" title="Movimientos"> <i class="bi bi-list-task"></i></a>
+                            <a href="<?= URL ?>cuentas/delete/<?= $cuenta->id ?>" title="Eliminar" onclick="return confirm('Confirmar eliminación Cuenta') " class="btn btn-danger" <?= (!in_array($_SESSION['id_rol'], $GLOBALS['clientes']['delete'])) ?
+                                    'disabled' : null ?>> <i class="bi bi-trash"></i> </a>
+                            <a href="<?= URL ?>cuentas/editar/<?= $cuenta->id ?>" title="Editar" class="btn btn-primary <?= (!in_array($_SESSION['id_rol'], $GLOBALS['clientes']['edit'])) ?
+									'disabled' : null ?>"> <i class="bi bi-pencil"></i> </a>
+                            <a href="<?= URL ?>cuentas/mostrar/<?= $cuenta->id ?>" title="Mostrar" class="btn btn-warning<?= (!in_array($_SESSION['id_rol'], $GLOBALS['clientes']['show'])) ?
+                                    'disabled' : null ?>"> <i class="bi bi-eye"></i> </a>
                         </td>
                     </tr>
                 <?php endforeach; ?>
             </tbody>
             <tfoot>
                 <tr>
-                    <td colspan="9">Nº Registros:
-                        <?= $this->cuentas->rowCount() ?>
-                    </td>
+                    <td colspan="9">Nº Registros: <?= $this->cuentas->rowCount() ?> </td>
                 </tr>
             </tfoot>
 
@@ -87,5 +70,4 @@
     <!-- Bootstrap JS y popper -->
     <?php require_once "template/partials/javascript.php" ?>
 </body>
-
 </html>
